requirements.txt
dash
dash-bootstrap-components
pandas
plotly
requests
python-dotenv
gunicorn
numpy
dash-auth